Structure: Restructure payments
