+++ src/components/general/payments/MoneyDiff/index.ts
@@ -1,0 +1,1 @@
+export * from './MoneyDiff.component';
