+++ src/screens/main/main.screen.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-param-reassign */
 import { SectionHeader } from "@moneyboy/components/general/lists/SectionHeader";
-import { MoneyDiff, MoneyDiffProps } from "@moneyboy/components/general/payments/moneyDiff";
+import { MoneyDiff, MoneyDiffProps } from "@moneyboy/components/general/payments/MoneyDiff";
 import { Content } from "@moneyboy/components/general/structure/content";
 import { ViewBase } from "@moneyboy/components/general/structure/viewBase";
 import variables from "@moneyboy/config/variables";
